+++ torch_points3d/core/base_conv/message_passing.py
@@ -17,7 +17,6 @@ from torch_geometric.nn import (
     knn,
 )
 from torch_geometric.data import Batch
-import torch_points_kernels as tp
 
 from torch_points3d.core.base_conv.base_conv import *
 from torch_points3d.core.common_modules import *