+++ torch_points3d/models/model_interface.py
@@ -26,7 +26,7 @@ class CheckpointInterface(ABC):
         pass
 
     @abstractmethod
-    def load_state_dict(self, state):
+    def load_state_dict(self, state, strict=False):
         pass