+++ torch_points3d/modules/PPNet/__init__.py
@@ -1,0 +1,2 @@
+from .blocks import *
+from .ops import *