+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- More general API for Minkowski with support for Bottleneck blocks and Squeeze and excite.
+
+## 1.1.1
+### Added
+
+- Teaser support for registration
+- Examples for using pretrained registration models
+
+### Changed
+
+- Moved to PyTorch 1.6 as officialy supported PyTorch version
+
+### Bug fix
+- Add `context = ssl._create_unverified_context()`, `data = urllib.request.urlopen(url, context=context)` within `download_ulr`, so ModelNet and ShapeNet can download.
+
+
 ## 1.1.0
 
 ### Added
@@ -24,6 +44,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Windows support
 - Weights are uploaded to wandb at the end of the run
 - Added PointGroup https://arxiv.org/pdf/2007.01294.pdf
+- Added PretrainedRegistry allowing model weight to be downloaded directly from wandb and DatasetMocking
+- Added script for s3dis cross-validation [scripts/cv_s3dis.py]. 6 different pretrained models will be downloaded, evaluated on full resolution and confusion matrice will be summed to get all metrics.
+- mAP tracker for Panoptic segmentation
 
 ### Changed
 
@@ -35,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bugfix
 
+- LR secheduler resume is broken for update on batch number #328
 - ElasticDistortion transform is now fully functional
 
 ### Removed